working rdp connection with xrdp; TODO verify this didnt break virtualbox compatibility

+++ nars_lab/src/main/java/automenta/rdp/Orders.java
@@ -1209,7 +1209,7 @@ public class Orders {
 
         if (delta) {
             change = (byte) data.get8();
-            coordinate += (int) change;
+            coordinate += change;
             return coordinate;
         } else {
             coordinate = data.getLittleEndian16();
